+++ client/src/components/AdminControls.scss
@@ -1,4 +1,4 @@
-@import "../../variables";
+@import "../variables";
 
 table {
   border-collapse: collapse;
