--- client/src/components/controls/EmployerControls.scss
@@ -1,1 +1,0 @@
-@import "../../_variables.scss";
